Split tmsh__commands.py into one file per command
Break the monolithic tmsh__commands.py (28 commands) into individual
files following the one-command-per-file convention used by iapp::*
commands. Move the shared _tmsh_spec helper into _base.py.
