+++ core/commands/registry/iapps/__init__.py
@@ -27,6 +27,40 @@ from . import (
     iapp__tmos_version,  # noqa: F401
     iapp__upgrade,  # noqa: F401
     iapp__upgrade_template,  # noqa: F401
+    script__help,  # noqa: F401
+    script__init,  # noqa: F401
+    script__run,  # noqa: F401
+    script__tabc,  # noqa: F401
+    tmsh__add_help,  # noqa: F401
+    tmsh__add_tabc,  # noqa: F401
+    tmsh__begin_transaction,  # noqa: F401
+    tmsh__builtin_help,  # noqa: F401
+    tmsh__builtin_tabc,  # noqa: F401
+    tmsh__cancel_transaction,  # noqa: F401
+    tmsh__cd,  # noqa: F401
+    tmsh__clear_screen,  # noqa: F401
+    tmsh__commit_transaction,  # noqa: F401
+    tmsh__create,  # noqa: F401
+    tmsh__delete,  # noqa: F401
+    tmsh__display,  # noqa: F401
+    tmsh__display_threshold,  # noqa: F401
+    tmsh__get_config,  # noqa: F401
+    tmsh__get_field_names,  # noqa: F401
+    tmsh__get_field_value,  # noqa: F401
+    tmsh__get_name,  # noqa: F401
+    tmsh__get_status,  # noqa: F401
+    tmsh__get_type,  # noqa: F401
+    tmsh__include,  # noqa: F401
+    tmsh__list,  # noqa: F401
+    tmsh__log,  # noqa: F401
+    tmsh__log_dest,  # noqa: F401
+    tmsh__log_level,  # noqa: F401
+    tmsh__modify,  # noqa: F401
+    tmsh__pwd,  # noqa: F401
+    tmsh__reset_stats,  # noqa: F401
+    tmsh__show,  # noqa: F401
+    tmsh__stateless,  # noqa: F401
+    tmsh__version,  # noqa: F401
 )
 from ._base import _REGISTRY
