+++ core/commands/registry/iapps/script__init.py
@@ -1,0 +1,20 @@
+"""script::init -- F5 TMSH script namespace command."""
+
+from __future__ import annotations
+
+from .._base import CommandDef
+from ..models import CommandSpec
+from ._base import _tmsh_spec, register
+
+
+@register
+class ScriptInit(CommandDef):
+    name = "script::init"
+
+    @classmethod
+    def spec(cls) -> CommandSpec:
+        return _tmsh_spec(
+            "script::init",
+            "Called before ``script::run``, ``script::help``, and ``script::tabc``.",
+            "script::init",
+        )
