+++ core/commands/registry/iapps/tmsh__get_field_names.py
@@ -1,0 +1,22 @@
+"""tmsh::get_field_names -- F5 TMSH object introspection command."""
+
+from __future__ import annotations
+
+from .._base import CommandDef
+from ..models import CommandSpec
+from ._base import _tmsh_spec, register
+
+
+@register
+class TmshGetFieldNames(CommandDef):
+    name = "tmsh::get_field_names"
+
+    @classmethod
+    def spec(cls) -> CommandSpec:
+        return _tmsh_spec(
+            "tmsh::get_field_names",
+            "Returns a list of field names present in an object.",
+            "tmsh::get_field_names <object>",
+            min_args=1,
+            max_args=1,
+        )
