+++ core/commands/registry/iapps/tmsh__list.py
@@ -1,0 +1,20 @@
+"""tmsh::list -- F5 TMSH configuration command."""
+
+from __future__ import annotations
+
+from .._base import CommandDef
+from ..models import CommandSpec
+from ._base import _tmsh_spec, register
+
+
+@register
+class TmshList(CommandDef):
+    name = "tmsh::list"
+
+    @classmethod
+    def spec(cls) -> CommandSpec:
+        return _tmsh_spec(
+            "tmsh::list",
+            "Runs the ``list`` command using the specified arguments.",
+            "tmsh::list ?component? ?name? ?options?",
+        )
